reviewsのカラム book_idのデータ型変更

+++ database/migrations/2020_03_05_121017_create_reviews_table.php
@@ -15,7 +15,7 @@ class CreateReviewsTable extends Migration
     {
         Schema::create('reviews', function (Blueprint $table) {
             $table->bigIncrements('user_id');
-            $table->bigIncrements('book_id');
+            $table->integer('book_id');
             $table->string('review');
             $table->string('practice');
             $table->timestamps();
